APIAccess
Initial project for accessing APIs. Google is currently the only supported API. A few design choices need to be figured out, and json output to html needs to be figured out, too.

A file with an example should come soon. API calls and returning entire results from YouTube can be done in a matter of a few lines. Many desired results take 4 or 5 separate calls to the YouTube API, so I may design a class which does these series of calls in a single function call.