APIAccess.php: Fixed session error handling
It turns out that I didn't properly handle errors on the API calls. Also, since the page is refreshed, the cURL settings weren't maintained after authorization.

Currently, session errors are handled during the first API call. If a call requires authorization but the credentials are invalid, the result of the json will contain an error. If any error is in the json, the session restarts. A class to handle specific errors appropriately may be considered.