Created APIAccess example
This example gets all the results of a playlist and returns them in an
array. The next step would be to find out the best way to display this
array to the user. A class dedicated to viewing the results will likely
be the next step, and this project will most likely follow the MVC
design pattern in future commits.